FDS-665 validation rules logic updated
